@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.2+3
+version: 1.0.3+4
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: '>=3.4.3 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,35 +29,35 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  google_fonts: ^4.0.3
+  google_fonts: ^6.2.1
   fluttericon: ^2.0.0
-  get: ^4.6.5
-  styled_widget: ^0.4.0+3
-  get_storage: ^2.0.3
-  audioplayers: ^3.0.1
+  get: ^4.6.6
+  styled_widget: ^0.4.1
+  get_storage: ^2.1.1
+  audioplayers: ^6.0.0
   enum_to_string: ^2.0.1
-  flutter_local_notifications: ^13.0.0
+  flutter_local_notifications: ^17.2.1+1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  bitsdojo_window: ^0.1.2
-  flutter_animate: ^3.1.0
-  dynamic_color: ^1.6.7
+  cupertino_icons: ^1.0.8
+  bitsdojo_window: ^0.1.6
+  flutter_animate: ^4.5.0
+  dynamic_color: ^1.7.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  msix: ^3.6.2
-  flutter_launcher_icons: ^0.11.0
-  change_app_package_name: ^1.1.0
+  msix: ^3.16.7
+  flutter_launcher_icons: ^0.13.1
+  change_app_package_name: ^1.3.0
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
