@@ -129,8 +129,8 @@ class _PomoState extends State<Pomo> {
       'main',
       channelDescription: 'Pomo Focus',
       subText: "Session running: ${phaseToString(pomoSession.currentPhase)}",
-      importance: Importance.min,
-      priority: Priority.min,
+      importance: Importance.max,
+      priority: Priority.max,
       onlyAlertOnce: true,
       when: pomoSession.endTimestamp,
       usesChronometer: true,
