@@ -17,14 +17,12 @@ main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
   // prevent landscape mode
-  SystemChrome.setPreferredOrientations(
-      [DeviceOrientation.portraitUp, DeviceOrientation.portraitDown]);
+  SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp, DeviceOrientation.portraitDown]);
 
-  await Future.delayed(const Duration(
-      milliseconds:
-          300)); // HACK: fix for https://github.com/flutter/flutter/issues/101007
-  if ([TargetPlatform.windows, TargetPlatform.linux, TargetPlatform.macOS]
-      .contains(defaultTargetPlatform)) {
+  await Future.delayed(
+    const Duration(milliseconds: 300),
+  ); // HACK: fix for https://github.com/flutter/flutter/issues/101007
+  if ([TargetPlatform.windows, TargetPlatform.linux, TargetPlatform.macOS].contains(defaultTargetPlatform)) {
     doWhenWindowReady(() {
       // appWindow.alignment = Alignment.center;
       appWindow.show();
@@ -65,8 +63,7 @@ const List<String> colorText = <String>[
   "Red",
 ];
 
-final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
-    FlutterLocalNotificationsPlugin();
+final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
 
 class MyApp extends StatefulWidget {
   const MyApp({Key? key}) : super(key: key);
@@ -95,12 +92,10 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> initNotifs() async {
-    FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
-        FlutterLocalNotificationsPlugin();
-    const AndroidInitializationSettings initializationSettingsAndroid =
-        AndroidInitializationSettings('launcher_icon');
-    const InitializationSettings initializationSettings =
-        InitializationSettings(
+    FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
+    const AndroidInitializationSettings initializationSettingsAndroid = AndroidInitializationSettings('launcher_icon');
+
+    const InitializationSettings initializationSettings = InitializationSettings(
       android: initializationSettingsAndroid,
     );
     await flutterLocalNotificationsPlugin.initialize(
@@ -108,9 +103,8 @@ class _MyAppState extends State<MyApp> {
     );
 
     flutterLocalNotificationsPlugin
-        .resolvePlatformSpecificImplementation<
-            AndroidFlutterLocalNotificationsPlugin>()
-        ?.requestPermission();
+        .resolvePlatformSpecificImplementation<AndroidFlutterLocalNotificationsPlugin>()
+        ?.requestNotificationsPermission();
   }
 
   void changeDarModeEnabled(bool newVal) {
@@ -163,11 +157,7 @@ class _MyAppState extends State<MyApp> {
         home: SafeArea(
           child: Column(
             children: [
-              if ([
-                TargetPlatform.windows,
-                TargetPlatform.linux,
-                TargetPlatform.macOS
-              ].contains(defaultTargetPlatform))
+              if ([TargetPlatform.windows, TargetPlatform.linux, TargetPlatform.macOS].contains(defaultTargetPlatform))
                 WindowTitleBarBox(
                   // the builder is needed for the context to find to the correct theme data
                   child: Builder(builder: (context) {
@@ -184,9 +174,7 @@ class _MyAppState extends State<MyApp> {
                             child: DefaultTextStyle(
                               style: TextStyle(
                                 fontSize: 12,
-                                color: darkModeEnabled
-                                    ? Colors.black
-                                    : Colors.white,
+                                color: darkModeEnabled ? Colors.black : Colors.white,
                                 fontWeight: FontWeight.normal,
                               ),
                               child: const Text(
@@ -231,12 +219,8 @@ class WindowButtons extends StatelessWidget {
             iconNormal: isDarkMode ? Colors.black : Colors.white,
             iconMouseDown: isDarkMode ? Colors.black : Colors.white,
             iconMouseOver: isDarkMode ? Colors.black : Colors.white,
-            mouseOver: isDarkMode
-                ? Colors.black.withOpacity(0.1)
-                : Colors.white.withOpacity(0.1),
-            mouseDown: isDarkMode
-                ? Colors.black.withOpacity(0.2)
-                : Colors.white.withOpacity(0.2),
+            mouseOver: isDarkMode ? Colors.black.withOpacity(0.1) : Colors.white.withOpacity(0.1),
+            mouseDown: isDarkMode ? Colors.black.withOpacity(0.2) : Colors.white.withOpacity(0.2),
           ),
         ),
         MaximizeWindowButton(
@@ -244,12 +228,8 @@ class WindowButtons extends StatelessWidget {
             iconNormal: isDarkMode ? Colors.black : Colors.white,
             iconMouseDown: isDarkMode ? Colors.black : Colors.white,
             iconMouseOver: isDarkMode ? Colors.black : Colors.white,
-            mouseOver: isDarkMode
-                ? Colors.black.withOpacity(0.1)
-                : Colors.white.withOpacity(0.1),
-            mouseDown: isDarkMode
-                ? Colors.black.withOpacity(0.2)
-                : Colors.white.withOpacity(0.2),
+            mouseOver: isDarkMode ? Colors.black.withOpacity(0.1) : Colors.white.withOpacity(0.1),
+            mouseDown: isDarkMode ? Colors.black.withOpacity(0.2) : Colors.white.withOpacity(0.2),
           ),
         ),
         CloseWindowButton(
@@ -308,8 +288,7 @@ class _MainState extends State<Main> {
     setState(() {
       _selectedIndex = index;
       pageChanged = true;
-      pageController.animateToPage(index,
-          duration: const Duration(milliseconds: 500), curve: Curves.easeInOut);
+      pageController.animateToPage(index, duration: const Duration(milliseconds: 500), curve: Curves.easeInOut);
     });
   }
 
@@ -350,23 +329,19 @@ class _MainState extends State<Main> {
             child: IconButton(
                 onPressed: () {
                   setState(() {
-                    Get.changeThemeMode(
-                        Get.isDarkMode ? ThemeMode.light : ThemeMode.dark);
+                    Get.changeThemeMode(Get.isDarkMode ? ThemeMode.light : ThemeMode.dark);
                     widget.changeDarModeEnabled(Get.isDarkMode);
                   });
                 },
-                icon: widget.darkModeEnabled
-                    ? const Icon(Icons.dark_mode, size: 18)
-                    : const Icon(Icons.sunny, size: 18)),
+                icon:
+                    widget.darkModeEnabled ? const Icon(Icons.dark_mode, size: 18) : const Icon(Icons.sunny, size: 18)),
           ),
           PopupMenuButton(
             tooltip: "Show color menu",
             icon: const Icon(Icons.color_lens),
             shape: RoundedRectangleBorder(
                 borderRadius: BorderRadius.circular(10),
-                side: BorderSide(
-                    color: Theme.of(context).colorScheme.secondaryContainer,
-                    width: 0.3)),
+                side: BorderSide(color: Theme.of(context).colorScheme.secondaryContainer, width: 0.3)),
             itemBuilder: (context) {
               return List.generate(colorOptions.length, (index) {
                 return PopupMenuItem(
@@ -376,15 +351,11 @@ class _MainState extends State<Main> {
                         Padding(
                           padding: const EdgeInsets.only(left: 10),
                           child: Icon(
-                            index == widget.colorSelected
-                                ? Icons.color_lens
-                                : Icons.color_lens_outlined,
+                            index == widget.colorSelected ? Icons.color_lens : Icons.color_lens_outlined,
                             color: colorOptions[index],
                           ),
                         ),
-                        Padding(
-                            padding: const EdgeInsets.only(left: 20),
-                            child: Text(colorText[index]))
+                        Padding(padding: const EdgeInsets.only(left: 20), child: Text(colorText[index]))
                       ],
                     ));
               });
